Move methods from BaseMetadataTest to SPITest
