+++ presto-connector-tests/src/main/java/com/facebook/presto/connector/unittest/BaseMetadataTest.java
@@ -15,27 +15,16 @@ package com.facebook.presto.connector.unittest;
 
 import com.facebook.presto.connector.meta.RequiredFeatures;
 import com.facebook.presto.spi.ColumnMetadata;
-import com.facebook.presto.spi.ConnectorOutputTableHandle;
 import com.facebook.presto.spi.ConnectorSession;
-import com.facebook.presto.spi.ConnectorTableHandle;
 import com.facebook.presto.spi.ConnectorTableMetadata;
 import com.facebook.presto.spi.SchemaTableName;
 import com.facebook.presto.spi.SchemaTablePrefix;
-import com.facebook.presto.spi.connector.Connector;
 import com.facebook.presto.spi.connector.ConnectorMetadata;
-import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
-import com.facebook.presto.spi.transaction.IsolationLevel;
 import com.facebook.presto.testing.TestingConnectorSession;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.io.Closer;
 import org.junit.jupiter.api.Test;
 
-import java.io.Closeable;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.Callable;
 import java.util.function.Consumer;
 
 import static com.facebook.presto.connector.meta.ConnectorFeature.CREATE_TABLE;
@@ -45,14 +34,11 @@ import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterables.getOnlyElement;
 import static java.util.Arrays.stream;
-import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public interface BaseMetadataTest
         extends SPITest
 {
-    Map<String, Object> getTableProperties();
-
     default List<String> systemSchemas()
     {
         return ImmutableList.of();
@@ -73,31 +59,6 @@ public interface BaseMetadataTest
     }
 
     /*
-     * Returns a SchemaTableName to be used when the name of the schema is NOT
-     * important.
-     *
-     * Tests that require tables in one schema should use this method. An
-     * example would be a test that renames a table within a single schema.
-     *
-     * Tests written using this can be run against connectors that support the
-     * creation of schemas and those that do not, but come with a usable schema
-     * out of the box. PostgreSQL, for example has a "public" schema.
-     *
-     * A class implementing this interface for e.g. PostgreSQL should return a
-     * SchemaTableName with the schemaName set to "public".
-     *
-     * Tests invoking this method do NOT generally need to be annotated with
-     * @RequiredFeatures({CREATE_SCHEMA, DROP_SCHEMA}).
-     *
-     * A class overriding this method should also override withSchemas, as
-     * described below.
-     */
-    default SchemaTableName schemaTableName(String tableName)
-    {
-        return new SchemaTableName("default_schema", tableName);
-    }
-
-    /*
      * Returns a SchemaTableName to be used when the name of the schema is
      * important.
      *
@@ -110,32 +71,6 @@ public interface BaseMetadataTest
     default SchemaTableName schemaTableName(String schemaName, String tableName)
     {
         return new SchemaTableName(schemaName, tableName);
-    }
-
-    /*
-     * Create and drop the schemas specified by schemaNames, running callable
-     * in between creating and dropping the schemas.
-     *
-     * A class implementing this interface for a connector that does NOT
-     * support creating and/or dropping schemas, but DOES have a usable schema
-     * (e.g. PostgreSQL with the "public" schema) should override this method
-     * and invoke callable without trying to create or destroy any of the
-     * schemas specified.
-     *
-     * Overriding this method in such a way allows tests that explicitly
-     * require creating and dropping tables to be written such that they are
-     * independent of whether or not the connector supports creating and
-     * dropping schemas.
-     */
-    default void withSchemas(ConnectorSession session, List<String> schemaNames, Callable<Void> callable)
-            throws Exception
-    {
-        try (Closer cleanup = closerOf(schemaNames.stream()
-                .distinct()
-                .map(schemaName -> new Schema(this, session, schemaName))
-                .collect(toImmutableList()))) {
-            callable.call();
-        }
     }
 
     @Test
@@ -187,22 +122,6 @@ public interface BaseMetadataTest
                         metadata -> metadata.dropTable(session, metadata.getTableHandle(session, schemaTableName))));
     }
 
-    default List<String> schemaNamesOf(SchemaTableName... schemaTableNames)
-    {
-        return stream(schemaTableNames)
-                .map(SchemaTableName::getSchemaName)
-                .collect(toImmutableList());
-    }
-
-    default List<String> schemaNamesOf(List<ConnectorTableMetadata> tables)
-    {
-        return tables.stream()
-                .map(ConnectorTableMetadata::getTable)
-                .map(SchemaTableName::getSchemaName)
-                .distinct()
-                .collect(toImmutableList());
-    }
-
     default SchemaTablePrefix prefixOfSchemaName(SchemaTableName schemaTableName)
     {
         return new SchemaTablePrefix(schemaTableName.getSchemaName());
@@ -218,36 +137,11 @@ public interface BaseMetadataTest
         consumers.forEach(this::withMetadata);
     }
 
-    default void withMetadata(Consumer<ConnectorMetadata> consumer)
+    default List<String> schemaNamesOf(SchemaTableName... schemaTableNames)
     {
-        Connector connector = getConnector();
-        ConnectorTransactionHandle transaction = connector.beginTransaction(IsolationLevel.READ_UNCOMMITTED, true);
-        ConnectorMetadata metadata = connector.getMetadata(transaction);
-        consumer.accept(metadata);
-        connector.commit(transaction);
-    }
-
-    class Schema
-            implements Closeable
-    {
-        private final BaseMetadataTest test;
-        private final ConnectorSession session;
-        private final String name;
-
-        public Schema(BaseMetadataTest test, ConnectorSession session, String name)
-        {
-            this.test = requireNonNull(test, "test is null");
-            this.session = requireNonNull(session, "session is null");
-            this.name = requireNonNull(name, "name is null");
-
-            test.withMetadata(metadata -> metadata.createSchema(session, name, ImmutableMap.of()));
-        }
-
-        @Override
-        public void close()
-        {
-            test.withMetadata(metadata -> metadata.dropSchema(session, name));
-        }
+        return stream(schemaTableNames)
+                .map(SchemaTableName::getSchemaName)
+                .collect(toImmutableList());
     }
 
     /*
@@ -263,52 +157,13 @@ public interface BaseMetadataTest
                 });
     }
 
-    class Table
-            implements Closeable
-    {
-        private final BaseMetadataTest test;
-        private final ConnectorSession session;
-        private final ConnectorTableMetadata tableMetadata;
-
-        public Table(BaseMetadataTest test, ConnectorSession session, ConnectorTableMetadata tableMetadata)
-        {
-            this.test = requireNonNull(test, "test is null");
-            this.session = requireNonNull(session, "session is null");
-            this.tableMetadata = requireNonNull(tableMetadata, "tableMetadata is null");
-
-            test.withMetadata(metadata -> {
-                ConnectorOutputTableHandle handle = metadata.beginCreateTable(session, tableMetadata, Optional.empty());
-                metadata.finishCreateTable(session, handle, ImmutableList.of());
-            });
-        }
-
-        @Override
-        public void close()
-        {
-            test.withMetadata(metadata -> {
-                ConnectorTableHandle handle = metadata.getTableHandle(session, tableMetadata.getTable());
-                metadata.dropTable(session, handle);
-            });
-        }
-    }
-
     default void withTables(ConnectorSession session, List<ConnectorTableMetadata> tables, List<Consumer<ConnectorMetadata>> consumers)
             throws Exception
     {
-        withSchemas(session, schemaNamesOf(tables), () -> {
-            try (Closer cleanup = closerOf(tables.stream()
-                    .map(table -> new Table(this, session, table))
-                    .collect(toImmutableList()))) {
-                consumers.forEach(this::withMetadata);
-            }
-            return null;
-        });
-    }
-
-    default Closer closerOf(List<Closeable> closables)
-    {
-        Closer result = Closer.create();
-        closables.forEach(result::register);
-        return result;
+        withTables(session, tables,
+                () -> {
+                    consumers.forEach(this::withMetadata);
+                    return null;
+                });
     }
 }
